requirements.txt
Flask==3.1.1
requests==2.32.4
python-dotenv==1.0.1
google-genai>=1.49.0
xai-sdk>=0.1.0

# Database & ORM
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
alembic==1.14.0
pyodbc==5.2.0

# Authentication & Security
Flask-Login==0.6.3
Flask-Limiter==3.8.0
Werkzeug==3.1.3
PyJWT==2.9.0
pyotp==2.9.0
qrcode==8.0
Pillow==10.4.0
cryptography==44.0.0

# RAG (Retrieval-Augmented Generation)
chromadb>=0.4.22
sentence-transformers>=2.2.2
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.2
tiktoken>=0.5.2

# Development Tools
pytest==8.3.4
black==24.10.0
flake8==7.1.1
mypy==1.13.0